requirements.txt
instagrapi
huggingface_hub
google-genai
Pillow
python-dotenv
requests
